New report print, level management and hier sort
Start of the Print report -- used to give a 1 page work sheet for a project
Changed level to be part of the object rather than part of the walk.
Level was need to support hier printing as a Format rather than
embeded in the Report hier.
Added hier sort to allow  ordering by Hierarcy.